pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vertixia-sdk"
version = "0.1.0"
description = "Official Python SDK for Vertixia AI OS - Build intelligent AI components and workflows"
authors = [
    {name = "Vertixia", email = "support@vertixia.ai"}
]
license = {text = "MIT"}
readme = "README.md"
homepage = "https://vertixia.ai"
repository = "https://github.com/vertixia/vertixia-sdk"
documentation = "https://docs.vertixia.ai/sdk"
keywords = ["ai", "automation", "agents", "workflows", "components"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "PyYAML>=6.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.24.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]
all = [
    "vertixia-sdk[dev,test,docs]"
]

[project.scripts]
vertixia = "vertixia_sdk.cli.main:main"

[project.urls]
"Bug Tracker" = "https://github.com/vertixia/vertixia-sdk/issues"
"Discussions" = "https://github.com/vertixia/vertixia-sdk/discussions"
"Changelog" = "https://github.com/vertixia/vertixia-sdk/blob/main/CHANGELOG.md"
"Documentation" = "https://docs.vertixia.ai/sdk"

[tool.hatch.build.targets.wheel]
packages = ["vertixia_sdk"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=vertixia_sdk --cov-report=term-missing"